fix another test

+++ backend/tests/integration/test_sciencedirect.py
@@ -33,7 +33,7 @@ def test_sciencedirect_response_returns_correct_elements():
         assert "color" in item
         assert item["color"] == "Not Relevant" or item["color"] == "SemiRelevant" or item["color"] == "Relevant"
         assert "relevance_score" in item
-        assert isinstance(item["relevance_score"], int)
+        assert isinstance(item["relevance_score"], float)
         assert item["relevance_score"] >= 0 & item["relevance_score"] <= 100
 
 def test_sciencedirect_student_rating_information_available():
